+++ resources/media/main.js
@@ -5,19 +5,8 @@
     const vscode = acquireVsCodeApi();
     const regexp = /scale\(([0-9\.]+)\)/;
 
-    // Handle messages sent from the extension to the webview
-    window.addEventListener('message', (event) => {
-        const message = event.data // The json data that the extension sent
-        switch (message.command) {
-            case 'refactor':
-                currentCount = Math.ceil(currentCount * 0.5)
-                counter.textContent = `${currentCount}`
-                break
-        }
-    })
-
     // Set VS Code state
-    let state = JSON.parse(document.getElementById('skin-panel-container').getAttribute('data-state'))
+    let state = JSON.parse(document.getElementById('skin-panel').getAttribute('data-state'))
     vscode.setState(state)
 
     window.addEventListener(
@@ -34,32 +23,48 @@
             document.getElementById('zoom_reset').addEventListener('click', zoomReset)
             document.getElementById('zoom_fit').addEventListener('click', zoomFit)
             document.getElementById('skin-panel-ddl').addEventListener('change', selectPanel)
-            var cont = document.getElementById('skin-panel-container')
+            document.getElementById('show_skin_border').addEventListener('click', toggleBorders)
+            var cont = document.getElementById('skin-panel')
             cont.style.transform = 'scale(' + (state.zoom === 'fit' ? calcFitScale() || 1 : state.zoom) + ')'
         },
         false
     )
     function selectPanel() {
-        var ddl = document.getElementById('skin-panel-ddl')
-        state.selectedPanelID = ddl.value
-        vscode.setState(state)
+        var ddl = document.getElementById('skin-panel-ddl');
+        state.selectedPanelID = ddl.value;
+        vscode.setState(state);
         vscode.postMessage({
             command: 'setState',
             body: state,
-        })
+        });
+    }
+    function toggleBorders() {
+        var cbx = document.getElementById('show_skin_border');
+        var cont = document.getElementById('skin-panel');
+        if (cbx.checked) {
+            cont.classList.add('display-borders');
+        } else {
+            cont.classList.remove('display-borders');
+        }
+        state.showBorders = cbx.checked;
+        vscode.setState(state);
+        vscode.postMessage({
+            command: 'setState',
+            body: state,
+        });
     }
     function zoomIn() {
-        var cont = document.getElementById('skin-panel-container')
-        var zoomFloat = currentZoomValue(cont)
-        zoomFloat += zoomFloat * 0.1
-        setZoom(cont, zoomFloat)
+        var cont = document.getElementById('skin-panel');
+        var zoomFloat = currentZoomValue(cont);
+        zoomFloat += zoomFloat * 0.1;
+        setZoom(cont, zoomFloat);
     }
     function zoomOut() {
-        var cont = document.getElementById('skin-panel-container')
-        var zoomFloat = currentZoomValue(cont)
-        zoomFloat -= zoomFloat * 0.1
+        var cont = document.getElementById('skin-panel');
+        var zoomFloat = currentZoomValue(cont);
+        zoomFloat -= zoomFloat * 0.1;
         if (zoomFloat < 0.1) return
-        setZoom(cont, zoomFloat)
+        setZoom(cont, zoomFloat);
     }
     function setZoom(cont, zoomFloat, isFit) {
         cont.style.transform = 'scale(' + zoomFloat + ')'
@@ -71,17 +76,17 @@
         })
     }
     function zoomReset() {
-        var cont = document.getElementById('skin-panel-container')
+        var cont = document.getElementById('skin-panel')
         setZoom(cont, 1.0)
     }
     function calcFitScale() {
-        var cont = document.getElementById('skin-panel-container')
-        var scale = Math.min((document.documentElement.clientWidth - 15) / cont.offsetWidth, (document.documentElement.clientHeight - 15) / cont.offsetHeight)
+        var cont = document.getElementById('skin-panel')
+        var scale = Math.min((document.documentElement.clientWidth - 25) / cont.offsetWidth, (document.documentElement.clientHeight - 60) / cont.offsetHeight)
         return scale
     }
     function zoomFit() {
         var scale = calcFitScale()
-        setZoom(document.getElementById('skin-panel-container'), scale, true)
+        setZoom(document.getElementById('skin-panel'), scale, true)
     }
     function currentZoomValue(cont) {
         if (cont.style.transform) {